SEDC-153: moved micex to package
