+++ sedc-core/src/main/resources/com/sedc/symbolload/micex-securities.xsd
